requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
deep-translator==1.11.4
idna==3.4
PySastrawi==1.2.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
